refactor: UserCard component to Staff component
